+++ lib/src/rotation_sensor_web_events_w3c.dart
@@ -10,6 +10,9 @@ import 'rotation_sensor_web_events.dart';
 
 /// A web implementation of the [RotationSensorPlatform].
 class RotationSensorWebEventsW3c extends RotationSensorWebEvents {
+  @override
+  RotationSensorWebEventsW3c newInstance() => RotationSensorWebEventsW3c();
+
   late final JSFunction _onDataJS = onData.toJS;
 
   String? _listeningEvent;