+++ lib/src/rotation_sensor_web_sensor_api.dart
@@ -12,6 +12,9 @@ import 'rotation_sensor_web.dart';
 
 /// A web implementation of the [RotationSensorPlatform].
 class RotationSensorWebSensorApi extends RotationSensorWeb {
+  @override
+  RotationSensorWebSensorApi newInstance() => RotationSensorWebSensorApi();
+
   late final JSFunction _onDataJS = _onData.toJS;
 
   late final JSFunction _onErrorJS = _onError.toJS;